add sample/wb-yoroi16

+++ include/T_tables.h
@@ -2,10 +2,10 @@
 #define T_TABLES_H
 
 // the T tables used in whitebox yoroi-16
-extern unsigned int T1[65536];
-extern unsigned int T2[65536];
-extern unsigned int T3[65536];
-extern unsigned int T1_inv[65536];
-extern unsigned int T2_inv[65536];
-extern unsigned int T3_inv[65536];
+extern unsigned short T1[65536];
+extern unsigned short T2[65536];
+extern unsigned short T3[65536];
+extern unsigned short T1_inv[65536];
+extern unsigned short T2_inv[65536];
+extern unsigned short T3_inv[65536];
 #endif
